HBP-90. Extract methods used only in HPC flow from OperationService to HPCOperationService.
