+++ framework_tvb/tvb/core/services/hpc_operation_service.py
@@ -1,0 +1,120 @@
+# -*- coding: utf-8 -*-
+#
+#
+# TheVirtualBrain-Framework Package. This package holds all Data Management, and
+# Web-UI helpful to run brain-simulations. To use it, you also need do download
+# TheVirtualBrain-Scientific Package (for simulators). See content of the
+# documentation-folder for more details. See also http://www.thevirtualbrain.org
+#
+# (c) 2012-2020, Baycrest Centre for Geriatric Care ("Baycrest") and others
+#
+# This program is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software Foundation,
+# either version 3 of the License, or (at your option) any later version.
+# This program is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+# PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+# You should have received a copy of the GNU General Public License along with this
+# program.  If not, see <http://www.gnu.org/licenses/>.
+#
+#
+#   CITATION:
+# When using The Virtual Brain for scientific publications, please cite it as follows:
+#
+#   Paula Sanz Leon, Stuart A. Knock, M. Marmaduke Woodman, Lia Domide,
+#   Jochen Mersmann, Anthony R. McIntosh, Viktor Jirsa (2013)
+#       The Virtual Brain: a simulator of primate brain network dynamics.
+#   Frontiers in Neuroinformatics (7:10. doi: 10.3389/fninf.2013.00010)
+#
+#
+
+"""
+.. moduleauthor:: Paula Popa <paula.popa@codemart.ro>
+.. moduleauthor:: Bogdan Valean <bogdan.valean@codemart.ro>
+"""
+import json
+import os
+from functools import partial
+
+from pyunicore.client import Job, Transport
+from tvb.basic.logger.builder import get_logger
+from tvb.core.entities.model.model_operation import STATUS_ERROR, STATUS_CANCELED, STATUS_FINISHED, STATUS_STARTED, \
+    STATUS_PENDING
+from tvb.core.entities.storage import dao
+from tvb.core.services.backend_clients.hpc_scheduler_client import HPCSchedulerClient, HPCJobStatus
+
+
+class HPCOperationService(object):
+    LOGGER = get_logger(__name__)
+
+    @staticmethod
+    def _operation_error(operation):
+        operation.mark_complete(STATUS_ERROR)
+        dao.store_entity(operation)
+
+    @staticmethod
+    def _operation_canceled(operation):
+        operation.mark_complete(STATUS_CANCELED)
+        dao.store_entity(operation)
+
+    @staticmethod
+    def _operation_started(operation):
+        operation.start_now()
+        dao.store_entity(operation)
+
+    @staticmethod
+    def _operation_finished(operation, simulator_gid):
+        op_ident = dao.get_operation_process_for_operation(operation.id)
+        # TODO: Handle login
+        job = Job(Transport(os.environ[HPCSchedulerClient.CSCS_LOGIN_TOKEN_ENV_KEY]),
+                  op_ident.job_id)
+        sim_h5_filenames, metric_op, metric_h5_filename = \
+            HPCSchedulerClient.stage_out_to_operation_folder(job.working_dir, operation, simulator_gid)
+        operation.mark_complete(STATUS_FINISHED)
+        dao.store_entity(operation)
+        HPCSchedulerClient().update_db_with_results(operation, sim_h5_filenames, metric_op, metric_h5_filename)
+
+    @staticmethod
+    def handle_hpc_status_changed(operation, simulator_gid, new_status):
+        # type: (Operation, str, str) -> None
+
+        switcher = {
+            STATUS_ERROR: HPCOperationService._operation_error,
+            STATUS_CANCELED: HPCOperationService._operation_canceled,
+            STATUS_STARTED: HPCOperationService._operation_started,
+            STATUS_FINISHED: partial(HPCOperationService._operation_finished, simulator_gid=simulator_gid)
+        }
+        update_func = switcher.get(new_status, lambda: "Invalid operation status")
+        update_func(operation)
+
+    @staticmethod
+    def check_operations_job():
+        operations = dao.get_operations()
+        if operations is None or len(operations) == 0:
+            HPCOperationService.LOGGER.info("There aren't any simulations in status PENDING or RUNNING")
+            return
+
+        for operation in operations:
+            try:
+                op_ident = dao.get_operation_process_for_operation(operation.id)
+                if op_ident is not None:
+                    transport = Transport(os.environ[HPCSchedulerClient.CSCS_LOGIN_TOKEN_ENV_KEY])
+                    job = Job(transport, op_ident.job_id)
+                    job_status = job.properties['status']
+                    if job.is_running():
+                        if operation.status == STATUS_PENDING and job_status == HPCJobStatus.READY.value:
+                            HPCOperationService._operation_started(operation)
+                        HPCOperationService.LOGGER.info(
+                            "CSCS job status: {} for operation {}.".format(job_status, operation.id))
+                        return
+                    HPCOperationService.LOGGER.info(
+                        "Job for operation {} has status {}".format(operation.id, job_status))
+                    if job_status == HPCJobStatus.SUCCESSFUL.value:
+                        simulator_gid = json.loads(operation.parameters)['gid']
+                        HPCOperationService._operation_finished(operation, simulator_gid)
+                    else:
+                        HPCOperationService._operation_error(operation)
+            except Exception:
+                HPCOperationService.LOGGER.error(
+                    "There was an error on background processing process for operation {}".format(operation.id),
+                    exc_info=True)
